changes Plane.polarization.dimension to long to match xsd and symmetry

+++ caom2/src/main/java/ca/nrc/cadc/caom2/Polarization.java
@@ -82,7 +82,7 @@ public class Polarization implements Serializable
 
     public List<PolarizationState> states;
 
-    public Integer dimension;
+    public Long dimension;
     
      @Override
     public String toString()
